GreystoneProject/requirements.txt
Django==1.11.3
djangorestframework==3.6.3
pytz==2017.2
requests==2.9.1